# Remove span builder