--- a/samples/Exporters/TestApplicationInsights.cs
+++ b/samples/Exporters/TestApplicationInsights.cs
@@ -65,7 +65,7 @@ namespace Samples
 
             using (tagContextBuilder.BuildScoped())
             {
-                using (tracer.WithSpan(tracer.CreateSpan("incoming request")))
+                using (tracer.WithSpan(tracer.StartSpan("incoming request")))
                 {
                     tracer.CurrentSpan.AddEvent("Start processing video.");
                     Thread.Sleep(TimeSpan.FromMilliseconds(10));