--- a/samples/Exporters/TestHttpClient.cs
+++ b/samples/Exporters/TestHttpClient.cs
@@ -42,7 +42,7 @@ namespace Samples
             var tracer = tracerFactory.GetTracer(nameof(HttpClientCollector));
             using (new HttpClientCollector(new HttpClientCollectorOptions(), tracer))
             {
-                using (tracer.WithSpan(tracer.CreateSpan("incoming request")))
+                using (tracer.WithSpan(tracer.StartSpan("incoming request")))
                 {
                     using (var client = new HttpClient())
                     {