--- a/samples/Exporters/TestJaeger.cs
+++ b/samples/Exporters/TestJaeger.cs
@@ -43,7 +43,7 @@ namespace Samples
             var tracer = tracerFactory.GetTracer(string.Empty);
 
             // Create a scoped span. It will end automatically when using statement ends
-            using (tracer.WithSpan(tracer.CreateSpan("Main")))
+            using (tracer.WithSpan(tracer.StartSpan("Main")))
             {
                 tracer.CurrentSpan.SetAttribute("custom-attribute", 55);
                 Console.WriteLine("About to do a busy work");
@@ -63,7 +63,7 @@ namespace Samples
             // Start another span. If another span was already started, it'll use that span as the parent span.
             // In this example, the main method already started a span, so that'll be the parent span, and this will be
             // a child span.
-            using (tracer.WithSpan(tracer.CreateSpan("DoWork")))
+            using (tracer.WithSpan(tracer.StartSpan("DoWork")))
             {
                 // Simulate some work.
                 var span = tracer.CurrentSpan;