--- a/samples/Exporters/TestRedis.cs
+++ b/samples/Exporters/TestRedis.cs
@@ -51,7 +51,7 @@ namespace Samples
             var db = connection.GetDatabase();
 
             // Create a scoped span. It will end automatically when using statement ends
-            using (tracer.WithSpan(tracer.SpanBuilder("Main").StartSpan()))
+            using (tracer.WithSpan(tracer.CreateSpan("Main")))
             {
                 Console.WriteLine("About to do a busy work");
                 for (var i = 0; i < 10; i++)
@@ -70,7 +70,7 @@ namespace Samples
             // Start another span. If another span was already started, it'll use that span as the parent span.
             // In this example, the main method already started a span, so that'll be the parent span, and this will be
             // a child span.
-            using (tracer.WithSpan(tracer.SpanBuilder("DoWork").StartSpan()))
+            using (tracer.WithSpan(tracer.CreateSpan("DoWork")))
             {
                 // Simulate some work.
                 var span = tracer.CurrentSpan;