--- a/samples/Exporters/TestStackdriver.cs
+++ b/samples/Exporters/TestStackdriver.cs
@@ -65,7 +65,7 @@ namespace Samples
 
             using (tagContextBuilder.BuildScoped())
             {
-                using (tracer.WithSpan(tracer.CreateSpan("incoming request")))
+                using (tracer.WithSpan(tracer.StartSpan("incoming request")))
                 {
                     tracer.CurrentSpan.AddEvent("Processing video.");
                     Thread.Sleep(TimeSpan.FromMilliseconds(10));