--- a/samples/LoggingTracer/LoggingTracer/LoggingTracer.cs
+++ b/samples/LoggingTracer/LoggingTracer/LoggingTracer.cs
@@ -2,13 +2,11 @@
 // Copyright (c) PlaceholderCompany. All rights reserved.
 // </copyright>
 
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-
 namespace LoggingTracer
 {
     using System;
+    using System.Collections.Generic;
+    using System.Diagnostics;
     using OpenTelemetry.Context.Propagation;
     using OpenTelemetry.Trace;
 
@@ -20,7 +18,7 @@ namespace LoggingTracer
         {
             Logger.Log("Tracer.ctor()");
         }
-        
+
         public ISpan CurrentSpan => CurrentSpanUtils.CurrentSpan;
 
         public IBinaryFormat BinaryFormat => new LoggingBinaryFormat();
@@ -33,69 +31,15 @@ namespace LoggingTracer
             return new CurrentSpanUtils.LoggingScope(span);
         }
 
-        public ISpan StartRootSpan(string operationName)
-        {
-            Logger.Log($"{prefix}.StartRootSpan({operationName})");
-            return new LoggingSpan(operationName, SpanKind.Internal);
-        }
-
-        public ISpan StartRootSpan(string operationName, SpanKind kind)
-        {
-            Logger.Log($"{prefix}.StartRootSpan({operationName} {kind})");
-            return new LoggingSpan(operationName, kind);
-        }
-
-        public ISpan StartRootSpan(string operationName, SpanKind kind, DateTimeOffset startTimestamp)
-        {
-            Logger.Log($"{prefix}.StartRootSpan({operationName}, {kind}, {startTimestamp:o})");
-            return new LoggingSpan(operationName, kind);
-        }
-
         public ISpan StartRootSpan(string operationName, SpanKind kind, DateTimeOffset startTimestamp, IEnumerable<Link> links)
         {
             Logger.Log($"{prefix}.StartRootSpan({operationName}, {kind}, {startTimestamp:o}, {links})");
             return new LoggingSpan(operationName, kind);
         }
 
-        public ISpan StartSpan(string operationName)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName})");
-            return new LoggingSpan(operationName, SpanKind.Internal);
-        }
-
-        public ISpan StartSpan(string operationName, SpanKind kind)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName} {kind})");
-            return new LoggingSpan(operationName, kind);
-        }
-
-        public ISpan StartSpan(string operationName, SpanKind kind, DateTimeOffset startTimestamp)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {kind}, {startTimestamp:o})");
-            return new LoggingSpan(operationName, kind);
-        }
-
         public ISpan StartSpan(string operationName, SpanKind kind, DateTimeOffset startTimestamp, IEnumerable<Link> links)
         {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {kind}, {startTimestamp:o} {links})");
-            return new LoggingSpan(operationName, kind);
-        }
-
-        public ISpan StartSpan(string operationName, ISpan parent)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name})");
-            return new LoggingSpan(operationName, SpanKind.Internal);
-        }
-
-        public ISpan StartSpan(string operationName, ISpan parent, SpanKind kind)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name}, {kind})");
-            return new LoggingSpan(operationName, kind);
-        }
-
-        public ISpan StartSpan(string operationName, ISpan parent, SpanKind kind, DateTimeOffset startTimestamp)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name}, {kind} {startTimestamp:o})");
+            Logger.Log($"{prefix}.StartSpan({operationName}, {kind}, {startTimestamp:o}, {links})");
             return new LoggingSpan(operationName, kind);
         }
 
@@ -105,46 +49,16 @@ namespace LoggingTracer
             return new LoggingSpan(operationName, kind);
         }
 
-        public ISpan StartSpan(string operationName, in SpanContext parent)
+        public ISpan StartSpan(string operationName, in SpanContext parent, SpanKind kind, DateTimeOffset startTimestamp, IEnumerable<Link> links)
         {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name})");
-            return new LoggingSpan(operationName, SpanKind.Internal);
-        }
-
-        public ISpan StartSpan(string operationName, in SpanContext parent, SpanKind kind)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name} {kind})");
+            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name}, {kind}, {startTimestamp:o}, {links})");
             return new LoggingSpan(operationName, kind);
         }
 
-        public ISpan StartSpan(string operationName, in SpanContext parent, SpanKind kind, DateTimeOffset startTimestamp)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name} {kind} {startTimestamp:o})");
-            return new LoggingSpan(operationName, kind);
-        }
-
-        public ISpan StartSpan(string operationName, in SpanContext parent, SpanKind kind, DateTimeOffset startTimestamp,
-            IEnumerable<Link> links)
-        {
-            Logger.Log($"{prefix}.StartSpan({operationName}, {parent.GetType().Name} {kind} {startTimestamp:o} {links})");
-            return new LoggingSpan(operationName, kind);
-        }
-
-        public ISpan StartSpanFromActivity(string operationName, Activity activity)
-        {
-            Logger.Log($"{prefix}.StartSpanFromActivity({operationName}, {activity.OperationName})");
-            return new LoggingSpan(operationName, SpanKind.Internal);
-        }
-
-        public ISpan StartSpanFromActivity(string operationName, Activity activity, SpanKind kind)
-        {
-            Logger.Log($"{prefix}.StartSpanFromActivity({operationName}, {activity.OperationName} {kind})");
-            return new LoggingSpan(operationName, kind);
-        }
 
         public ISpan StartSpanFromActivity(string operationName, Activity activity, SpanKind kind, IEnumerable<Link> links)
         {
-            Logger.Log($"{prefix}.StartSpanFromActivity({operationName}, {activity.OperationName} {kind} {links})");
+            Logger.Log($"{prefix}.StartSpanFromActivity({operationName}, {activity.OperationName}, {kind}, {links})");
             return new LoggingSpan(operationName, kind);
         }
     }