--- a/src/OpenTelemetry.Abstractions/Trace/ISpan.cs
+++ b/src/OpenTelemetry.Abstractions/Trace/ISpan.cs
@@ -44,12 +44,6 @@ namespace OpenTelemetry.Trace
         Status Status { set; }
 
         /// <summary>
-        /// Gets a value indicating whether this span was already stopped.
-        /// </summary>
-        /// <remarks>This method is not compliant with the specification. https://github.com/open-telemetry/opentelemetry-specification/issues/55 .</remarks>
-        bool HasEnded { get; }
-
-        /// <summary>
         /// Updates the <see cref="ISpan"/> name.
         ///
         /// If used, this will override the name provided via CreateSpan method overload.