--- a/src/OpenTelemetry.Abstractions/Trace/ISpan.cs
+++ b/src/OpenTelemetry.Abstractions/Trace/ISpan.cs
@@ -52,7 +52,7 @@ namespace OpenTelemetry.Trace
         /// <summary>
         /// Updates the <see cref="ISpan"/> name.
         ///
-        /// If used, this will override the name provided via <see cref="ISpanBuilder"/>.
+        /// If used, this will override the name provided via CreateSpan method overload.
         /// Upon this update, any sampling behavior based on <see cref="ISpan"/> name will depend on the
         /// implementation.
         /// </summary>