… README.md
… benchmarks/Tracing/SpanCreationScenarios.cs
… samples/LoggingTracer/LoggingTracer/LoggingSpanBuilder.cs
… src/OpenTelemetry.Abstractions/Trace/ISpanBuilder.cs
… src/OpenTelemetry.Abstractions/Trace/ITracer.cs
… src/OpenTelemetry.Abstractions/Trace/NoopSpanBuilder.cs
… src/OpenTelemetry.Abstractions/Trace/NoopTracer.cs
… src/OpenTelemetry.Abstractions/Trace/TracerExtensions.cs
… src/OpenTelemetry.Collector.AspNetCore/Implementation/HttpInListener.cs
… src/OpenTelemetry.Collector.Dependencies/Implementation/AzureSdkDiagnosticListener.cs
… src/OpenTelemetry.Collector.Dependencies/Implementation/HttpHandlerDiagnosticListener.cs
… src/OpenTelemetry.Collector.StackExchangeRedis/Implementation/RedisProfilerEntryToSpanConverter.cs
… src/OpenTelemetry.Collector.StackExchangeRedis/OpenTelemetry.Collector.StackExchangeRedis.csproj
… src/OpenTelemetry.Collector.StackExchangeRedis/StackExchangeRedisCallsCollector.cs
… src/OpenTelemetry.Shims.OpenTracing/SpanBuilderShim.cs
… src/OpenTelemetry/Trace/Span.cs
… src/OpenTelemetry/Trace/SpanBuilder.cs
… src/OpenTelemetry/Trace/Tracer.cs
… test/OpenTelemetry.Collector.Dependencies.Tests/BasicTests.cs
… test/OpenTelemetry.Exporter.ApplicationInsights.Tests/Implementation/ApplicationInsightsTraceExporterTests.cs
… test/OpenTelemetry.Exporter.Jaeger.Tests/Implementation/JaegerSpanConverterTest.cs
… test/OpenTelemetry.Exporter.Jaeger.Tests/Implementation/JaegerThriftIntegrationTest.cs
… test/OpenTelemetry.Exporter.LightStep.Tests/LightStepSpanConverterTest.cs
… test/OpenTelemetry.Shims.OpenTracing.Tests/SpanBuilderShimTests.cs
… test/OpenTelemetry.Tests/Impl/Trace/CurrentSpanUtilsTest.cs
… test/OpenTelemetry.Tests/Impl/Trace/Export/BatchingSpanProcessorTests.cs
… test/OpenTelemetry.Tests/Impl/Trace/Export/SimpleSpanProcessorTests.cs
… test/OpenTelemetry.Tests/Impl/Trace/NoopSpanBuilderTests.cs
… test/OpenTelemetry.Tests/Impl/Trace/NoopTracerTests.cs
… test/OpenTelemetry.Tests/Impl/Trace/SpanBuilderTest.cs