Fixes #145 

SpanBuilder requires extra allocation for every Span created, it also makes span creation very verbose.

This PR suggest how to remove builder and create span directly with a lot of overloads on Tracer.

**Before**
```csharp
var span = tracer
    .SpanBuilder("outgoing http request")
    .SetSpanKind(SpanKind.Client)
    .StartSpan();
```

or 

```csharp
    var span = tracer
        .SpanBuilder("GET api/values") 
        .SetCreateChild(false) 
        .StartSpan();
```

**After**

```csharp
var span = tracer.StartSpan("outgoing http request", SpanKind.Client);
```

```csharp
var span = tracer.StartSpanFromActivity("GET api/values", Activity.Current)
```

## Benchmarks
- Noop allocations are gone, time to create noop span got twice as small
- for real span creation, there a slight decrease in allocations and time

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18362
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview8-013656
  [Host]     : .NET Core 2.2.6 (CoreCLR 4.6.27817.03, CoreFX 4.6.27818.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.6 (CoreCLR 4.6.27817.03, CoreFX 4.6.27818.02), 64bit RyuJIT

### Before
|                           Method |        Mean |      Error |     StdDev |  Ratio | RatioSD |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|--------------------------------- |------------:|-----------:|-----------:|-------:|--------:|-------:|-------:|-------:|----------:|
|               CreateSpan_Sampled |   835.12 ns | 15.7607 ns | 13.1609 ns |  67.59 |    3.12 | 0.1984 |      - |      - |     840 B |
|    CreateSpan_Attributes_Sampled | 1,054.59 ns | 13.8295 ns | 12.9361 ns |  85.64 |    3.59 | 0.3624 |      - |      - |    1528 B |
|     CreateSpan_Propagate_Sampled | 1,893.55 ns | 31.8505 ns | 29.7930 ns | 153.75 |    6.36 | 0.2804 | 0.0038 | 0.0019 |    1207 B |
| CreateSpan_Attributes_NotSampled |   871.99 ns | 17.2950 ns | 33.7326 ns |  72.29 |    4.57 | 0.1993 |      - |      - |     840 B |
|                  CreateSpan_Noop |    12.16 ns |  0.2692 ns |  0.4269 ns |   1.00 |    0.00 | 0.0057 |      - |      - |      24 B |
|       CreateSpan_Attributes_Noop |    22.32 ns |  0.3673 ns |  0.3436 ns |   1.81 |    0.07 | 0.0057 |      - |      - |      24 B |
|        CreateSpan_Propagate_Noop |    14.79 ns |  0.2758 ns |  0.2580 ns |   1.20 |    0.05 | 0.0057 |      - |      - |      24 B |

### After
|                           Method |         Mean |      Error |     StdDev |  Ratio | RatioSD |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|--------------------------------- |-------------:|-----------:|-----------:|-------:|--------:|-------:|-------:|-------:|----------:|
|               CreateSpan_Sampled |   803.906 ns | 15.9024 ns | 16.3306 ns | 131.65 |    3.78 | 0.1402 |      - |      - |     592 B |
|    CreateSpan_Attributes_Sampled | 1,032.477 ns | 25.2346 ns | 36.9885 ns | 171.31 |    6.92 | 0.3033 |      - |      - |    1280 B |
|     CreateSpan_Propagate_Sampled | 1,800.736 ns | 42.5193 ns | 53.7733 ns | 297.26 |   12.51 | 0.2232 | 0.0057 | 0.0038 |     972 B |
| CreateSpan_Attributes_NotSampled |   800.006 ns |  8.5102 ns |  7.9605 ns | 131.25 |    2.95 | 0.1402 |      - |      - |     592 B |
|                  CreateSpan_Noop |     6.097 ns |  0.1240 ns |  0.1160 ns |   1.00 |    0.00 |      - |      - |      - |         - |
|       CreateSpan_Attributes_Noop |    15.322 ns |  0.3289 ns |  0.5217 ns |   2.52 |    0.10 |      - |      - |      - |         - |
|        CreateSpan_Propagate_Noop |     7.787 ns |  0.0833 ns |  0.0779 ns |   1.28 |    0.03 |      - |      - |      - |         - |