@bogdandrutu 2019-10-10
Alternative option is to have an "StartOption" class for the optional fields. So you do have the "required" fields like name explicit and everything else in the Option. This will reduce the API surface and will achieve the performance benefit for the 90% of the cases.