@lmolkova 2019-10-10
> Alternative option is to have an "StartOption" class for the optional fields. So you do have the "required" fields like name explicit and everything else in the Option. This will reduce the API surface and will achieve the performance benefit for the 90% of the cases.

The problem with small class is the same as builder - it's an allocation on heap. Which is exactly we what to avoid by removing builder.

The option is to have struct Options (stack allocation), but this struct will be too big and compicates usage. Personally I like how this API looks now from usage perspective (not internally though but this is lesser priority)