@lmolkova 2019-10-11
> Overall in favor, but we should be aware of [open-telemetry/opentelemetry-specification#238](https://github.com/open-telemetry/opentelemetry-specification/issues/238) and [open-telemetry/opentelemetry-specification#287](https://github.com/open-telemetry/opentelemetry-specification/pull/287); `StartSpan` shouldn't necessarily activate the span.

@austinlparker  Thanks for sharing! I've gone through the issues and there seems to be a consensus that there is start and activate (make current). It seems there are no plans to separate *create* and *start*. Did I miss anything? 

Just to clarify: we still have start and tracer.WithSpan separate. This PR merges create + start. And if someone wants delayed start, they could also delay create in the same way.